requirements.txt
# requirements.txt

# Core dependencies
gradio==3.32.0
firebase-admin==6.1.0
google-cloud-storage==2.9.0
Pillow==9.5.0
requests==2.31.0
python-dotenv==1.0.0

# Additional dependencies that might be required
numpy==1.24.3


# If you're using any machine learning models locally
# torch==2.0.1
# torchvision==0.15.2

# For working with JSON data
json5==0.9.14

# For handling date and time
python-dateutil==2.8.2

# For better logging
colorlog==6.7.0

# For API interactions
httpx==0.24.1

# For working with SVG files (if needed)
svglib==1.5.1

# For working with environment variables
python-decouple==3.8

# For enhanced HTTP capabilities
aiohttp==3.8.4


# For working with Excel files (if needed)
openpyxl==3.1.2
itsdangerous

# New dependencies based on the provided imports
fastapi==0.95.2
starlette==0.27.0
python-multipart==0.0.6
